package.json: update react-data-grid to 5.0.3
Fixes an issue with Chrome where horizontal scrolling is broken.
Updating react-data-grid to a version that introduces scrolling fixes,
which is `5.0.1` will solve many scrolling issues, however npm resolves
to `5.0.3` so we'll bump to `5.0.3`.

We also include a dependency to 'immutable' because upstream relies on
it as a dev dependency, but when we build we must also require it as
well or the build will break.  For some reason npm doesn't do a good
enough job at figuring out transitive dependencies.

Signed-off-by: Jamie Couture <jamie@quandl.com>