datatable: introduce auto-width flag
Currently, DataTable provides a way for users to define column widths,
which is useful for flexibility.  However, when tables and pre-set
columns do not match up perfectly, the last column can fall short of the
calculated width of the table viewport.  Adding a new flag to allow
callers to skip setting the fixed width of the last column, such that it
will be calculated by react-table-grid instead, allowing it to be a sane
width.

Fixes an issue where setting fixed widths may cause last item to fall
short of calculated container.

Signed-off-by: Jamie Couture <jamie@quandl.com>